zadanie_9/requirements.txt
Flask-Cors~=5.0.0
openai~=1.44.1
PyYAML~=6.0.2
Flask~=3.0.3
